Remove BraveTorClientUpdaterDelegate and avoid unused variable warning
