+++ browser/tor/BUILD.gn
@@ -13,8 +13,6 @@ source_set("tor") {
 
   if (enable_tor) {
     sources = [
-      "brave_tor_client_updater_delegate.cc",
-      "brave_tor_client_updater_delegate.h",
       "onion_location_navigation_throttle_delegate.cc",
       "onion_location_navigation_throttle_delegate.h",
       "tor_profile_service_factory.cc",
